// src/types/beans.ts
export interface ICollector {
  id: number;
  ip: string;
  no: string;
  mac: string;
  pwd: string;
  createDt: string;
  off: boolean;
  extend: Record<string, any>;
}
export interface ICollectorWarn {
  id: number;
  collector: ICollector;
  info: string;
  createDt: string;
}
export interface IDepartmentRelation {
  department: IDepartment;
  users: number[];
  products: number[];
  callUsers: number[];
  userList: IUser[];
  productList: IProduct[];
  callUserList: IUser[];
}
export interface IDevice {
  id: string;
  collector: ICollector;
  name: string;
  type: string;
  createDt: string;
  off: boolean;
  extend: Record<string, any>;
  deviceType: IDeviceType;
}
// ...
export interface IDeviceType {
  id: string;
  name: string;
  img: string;
  extend: Record<string, any>;
}